refactor: types 세그먼트 제거 및 model types로 이동
