+++ apps/web/src/entities/notification/config/query/notificationQueryOptions.ts
@@ -7,7 +7,7 @@ import {
   getNotifications,
   getNotificationUnreadCnt,
 } from '../../api';
-import type { GetNotificationsQuery } from '../../types';
+import { type GetNotificationsQuery } from '../../model';
 
 import { notificationQueryKeys } from './notificationQueryKeys';
