+++ apps/web/src/entities/notification/index.ts
@@ -1,10 +1,5 @@
 export * from './model';
 export * from './lib';
-export type {
-  GetNotificationsQuery,
-  GetNotificationsResponseDto,
-  PatchNotificationReadStatusParam,
-} from './types';
 export { notificationHandler } from './mock';
 export {
   notificationQueryKeys,
