+++ apps/web/src/entities/notification/model/types/index.ts
@@ -1,1 +1,4 @@
 export * from './types';
+export type { GetNotificationsQuery } from './query';
+export type { GetNotificationsResponseDto } from './dto';
+export type { PatchNotificationReadStatusParam } from './param';
