+++ apps/web/src/entities/notification/model/types/param/patchNotificationReadStatusParam.ts
