+++ apps/web/src/entities/notification/model/types/query/getNotificationsQuery.ts
