--- apps/web/src/entities/notification/types/index.ts
@@ -1,3 +1,0 @@
-export type { GetNotificationsQuery } from './query';
-export type { GetNotificationsResponseDto } from './dto';
-export type { PatchNotificationReadStatusParam } from './param';
